refactor(ts-client): split Seismic tx adapters out of chain.ts
Extract Seismic tx typing, serialization, and RPC formatting out of chain.ts so the Seismic/viem adapter boundary is explicit.

Changes:
- add src/tx/seismicTx.ts for canonical Seismic tx/request types, SEISMIC_TX_TYPE, and serializeSeismicTransaction
- add src/tx/seismicRpc.ts for Seismic RPC schemas, RPC request formatting, and auth-list contractAddress -> address translation
- reduce src/chain.ts to chain definitions and formatter wiring
- stop re-exporting Seismic tx internals from chain.ts
- export Seismic tx types/serializer and RPC helpers directly from the new tx/ modules via the package root
- update internal imports to depend on tx/seismicTx.ts / tx/seismicRpc.ts directly instead of routing through chain.ts
- add top-of-file comments clarifying the responsibility of chain.ts, seismicTx.ts, and seismicRpc.ts

Rationale:
chain.ts had accumulated too many responsibilities: chain definitions, Seismic tx types, Seismic RLP serialization, and Seismic RPC formatting. Pulling the tx adapter logic into dedicated modules makes the architecture more honest and easier to trace, especially around Seismic/viem auth-list translation and future tx-type changes.
