+++ clients/ts/packages/seismic-viem/src/chain.ts
@@ -1,417 +1,9 @@
-import {
-  concatHex,
-  defineChain,
-  formatTransactionRequest,
-  hexToBigInt,
-  toHex,
-  toRlp,
-} from 'viem'
-import type {
-  Address,
-  BlockIdentifier,
-  BlockNumber,
-  BlockTag,
-  Chain,
-  ChainFormatters,
-  ExactPartial,
-  Hex,
-  OneOf,
-  RpcSchema,
-  RpcStateOverride,
-  SerializeTransactionFn,
-  Signature,
-  TransactionRequest,
-  TransactionRequestEIP7702,
-  TransactionRequestLegacy,
-  TransactionSerializable,
-  TransactionSerializableEIP7702,
-  TransactionSerializableLegacy,
-} from 'viem'
+// Seismic chain definitions only. Seismic tx typing/serialization lives under
+// `src/tx/`, and this module wires those formatters into the exported chains.
+import { defineChain } from 'viem'
+import type { Chain } from 'viem'
 
-import { toYParitySignatureArray } from '@sviem/viem-internal/signature.ts'
-
-export const SEISMIC_TX_TYPE = 74 // '0x4a'
-
-/**
- * The additional fields added to a Seismic transaction
- * @interface SeismicTxExtras
- * @property {Hex} [encryptionPubkey] - The public key used to encrypt the calldata. This uses AES encryption, where the user's keypair is combined with the network's keypair
- * @property {number} [messageVersion] - The version of the message being sent. Used for signing transactions via messages. Normal transactions use messageVersion = 0. Txs signed with EIP-712 use messageVersion = 2
- */
-
-type SeismicTxExtrasBlank = {
-  encryptionPubkey?: undefined
-  encryptionNonce?: undefined
-  messageVersion?: undefined
-  recentBlockHash?: undefined
-  expiresAtBlock?: undefined
-  signedRead?: undefined
-}
-
-export type SeismicTxExtras = {
-  encryptionPubkey?: Hex | undefined
-  encryptionNonce?: Hex | undefined
-  messageVersion?: number | undefined
-  recentBlockHash?: Hex | undefined
-  expiresAtBlock?: bigint | undefined
-  signedRead?: boolean | undefined
-}
-
-export type SeismicBlockParams = {
-  recentBlockHash: Hex
-  expiresAtBlock: bigint
-}
-
-export type SeismicElements = {
-  encryptionPubkey: Hex
-  encryptionNonce: Hex
-  messageVersion: number
-  signedRead: boolean
-} & SeismicBlockParams
-
-/**
- * Advanced overrides for Seismic metadata generation.
- *
- * Most callers should omit these and let the client derive safe defaults from
- * chain state. These are primarily useful for:
- *
- * - deterministic tests and debugging
- * - reproducing an exact encrypted or signed request shape
- * - explicit expiry-window control
- * - low-level interoperability with other Seismic tooling
- *
- * These options are only meaningful on Seismic signed or shielded paths. They
- * do not apply to transparent reads or writes.
- */
-export type SeismicSecurityParams = {
-  /**
-   * Expiry window used when deriving `expiresAtBlock` from a recent block.
-   * Defaults to the standard 100-block window.
-   * Ignored when `expiresAtBlock` is provided explicitly.
-   */
-  blocksWindow?: bigint
-  /**
-   * Override the AEAD encryption nonce used for the Seismic payload.
-   * Mainly useful for deterministic tests or reproducing an exact request.
-   * When omitted, a random nonce is generated.
-   */
-  encryptionNonce?: Hex
-  /**
-   * Override the recent block hash used for freshness and replay protection.
-   * If provided without `expiresAtBlock`, expiry is derived from this block plus
-   * `blocksWindow`.
-   * When omitted, the latest block hash is used.
-   */
-  recentBlockHash?: Hex
-  /**
-   * Override the block height after which the signed or shielded request is no
-   * longer valid.
-   * When omitted, expiry is derived from the chosen recent block plus
-   * `blocksWindow`.
-   */
-  expiresAtBlock?: bigint
-}
-
-/**
- * Represents a Seismic transaction request, extending viem's base {@link https://viem.sh/docs/glossary/types#transactionrequest TransactionRequest} with {@link SeismicTxExtras}
- *
- * @interface SeismicTransactionRequest
- * @extends {TransactionRequest}
- * @extends {SeismicTxExtras}
- */
-export type SeismicTransactionRequest =
-  | (TransactionRequest & SeismicTxExtrasBlank)
-  | (Omit<TransactionRequestLegacy, 'type'> &
-      SeismicTxExtras & {
-        authorizationList?: TransactionRequestEIP7702['authorizationList']
-        type: 'seismic'
-      })
-
-/**
- * Represents a serializable Seismic transaction, extending viem's base {@link https://viem.sh/docs/utilities/parseTransaction#returns TransactionSerializable} with {@link SeismicTxExtras}
- *
- * @interface TransactionSerializableSeismic
- * @extends {TransactionSerializable}
- * @extends {SeismicTxExtras}
- */
-export type TransactionSerializableSeismic =
-  | (TransactionSerializable & SeismicTxExtrasBlank)
-  | (Omit<TransactionSerializableLegacy, 'type'> &
-      SeismicTxExtras & {
-        authorizationList?: TransactionSerializableEIP7702['authorizationList']
-        type: 'seismic'
-      })
-
-export type TxSeismic = {
-  chainId?: number
-  nonce?: bigint
-  gasPrice?: bigint
-  gasLimit?: bigint
-  to?: Address | null
-  isCreate?: boolean
-  value?: bigint
-  input?: Hex
-  encryptionPubkey: Hex
-  encryptionNonce: Hex
-  messageVersion: number
-  recentBlockHash: Hex
-  expiresAtBlock: bigint
-  signedRead: boolean
-  authorizationList?: {
-    chainId: bigint
-    contractAddress: `0x${string}`
-    nonce: bigint
-    yParity: number
-    r: `0x${string}`
-    s: `0x${string}`
-  }[]
-}
-
-export type SeismicTxSerializer = SerializeTransactionFn<
-  TransactionSerializableSeismic,
-  'seismic'
->
-
-export const serializeSeismicTransaction: SeismicTxSerializer = (
-  tx: OneOf<TransactionSerializable | TransactionSerializableSeismic>,
-  signature?: Signature
-): Hex => {
-  const {
-    chainId,
-    nonce,
-    gasPrice,
-    gas,
-    to,
-    data,
-    value = 0n,
-    encryptionPubkey,
-    encryptionNonce,
-    messageVersion = 0,
-    recentBlockHash,
-    expiresAtBlock,
-    signedRead = false,
-    authorizationList,
-  } = tx
-
-  if (chainId === undefined) {
-    throw new Error('Seismic transactions require chainId')
-  }
-  if (nonce === undefined) {
-    throw new Error('Seismic transactions require nonce')
-  }
-  if (gasPrice === undefined) {
-    throw new Error('Seismic transactions require gasPrice')
-  }
-  if (gas === undefined) {
-    throw new Error('Seismic transactions require gas')
-  }
-  if (to === undefined) {
-    throw new Error('Seismic transactions require to')
-  }
-  // value, input can be undefined
-  if (encryptionPubkey === undefined) {
-    throw new Error('Seismic transactions require encryptionPubkey')
-  }
-  if (encryptionNonce === undefined) {
-    throw new Error('Seismic transactions require encryptionNonce')
-  }
-  if (recentBlockHash === undefined) {
-    throw new Error('Seismic transactions require recentBlockHash')
-  }
-  if (expiresAtBlock === undefined) {
-    throw new Error('Seismic transactions require expiresAtBlock')
-  }
-  if (data === undefined) {
-    throw new Error('Seismic transactions require input')
-  }
-
-  // Seismic elements are encoded FLAT (not nested) - each field is a separate RLP item
-  // Note: Zero values are encoded as empty bytes (0x) to match Rust's alloy-rlp encoding
-  const rlpArray = [
-    toHex(chainId),
-    nonce ? toHex(nonce) : '0x',
-    gasPrice ? toHex(gasPrice) : '0x',
-    gas ? toHex(gas) : '0x',
-    to ?? '0x',
-    value ? toHex(value) : '0x',
-    // Seismic elements fields (encoded flat, not as a nested list)
-    encryptionPubkey ?? '0x',
-    hexToBigInt(encryptionNonce) === 0n ? '0x' : encryptionNonce,
-    messageVersion === 0 ? '0x' : toHex(messageVersion),
-    recentBlockHash,
-    toHex(expiresAtBlock),
-    signedRead ? '0x01' : '0x',
-    // Input field comes after seismic elements
-    data ?? '0x',
-    // Authorization list: nested RLP list [[chainId, address, nonce, yParity, r, s], ...]
-    // Cast needed: viem's TransactionSerializable union sets authorizationList to undefined
-    // on most variants, collapsing the item type to never when destructured from OneOf<>.
-    ((authorizationList ?? []) as TxSeismic['authorizationList'] & []).map(
-      (auth) => [
-        auth.chainId ? toHex(auth.chainId) : '0x',
-        auth.contractAddress,
-        auth.nonce ? toHex(auth.nonce) : '0x',
-        auth.yParity ? toHex(auth.yParity) : '0x',
-        auth.r,
-        auth.s,
-      ]
-    ),
-    ...toYParitySignatureArray(tx as TransactionSerializableLegacy, signature),
-  ]
-
-  const rlpEncoded = toRlp(rlpArray as any)
-  return concatHex([toHex(SEISMIC_TX_TYPE), rlpEncoded])
-}
-
-export const estimateGasRpcSchema = {
-  Method: 'eth_estimateGas',
-  Parameters: ['SeismicTransactionRequest'] as
-    | [SeismicTransactionRequest]
-    | [SeismicTransactionRequest, BlockNumber]
-    | [SeismicTransactionRequest, BlockNumber, RpcStateOverride],
-  ReturnType: 'Quantity',
-}
-
-export const callRpcSchema = {
-  Method: 'eth_call',
-  Parameters: ['SeismicTransactionRequest'] as
-    | [ExactPartial<SeismicTransactionRequest>]
-    | [
-        ExactPartial<SeismicTransactionRequest>,
-        BlockNumber | BlockTag | BlockIdentifier,
-      ]
-    | [
-        ExactPartial<SeismicTransactionRequest>,
-        BlockNumber | BlockTag | BlockIdentifier,
-        RpcStateOverride,
-      ],
-  ReturnType: 'Hex',
-}
-
-export const seismicRpcSchema: RpcSchema = [estimateGasRpcSchema, callRpcSchema]
-
-const hasSeismicFields = (request: SeismicTransactionRequest) => {
-  return (
-    request.encryptionPubkey !== undefined &&
-    request.encryptionNonce !== undefined &&
-    request.messageVersion !== undefined &&
-    request.recentBlockHash !== undefined &&
-    request.expiresAtBlock !== undefined &&
-    request.signedRead !== undefined
-  )
-}
-
-const formatAuthorizationList = (
-  authorizationList: NonNullable<TransactionRequestEIP7702['authorizationList']>
-) =>
-  authorizationList.map((authorization) => ({
-    address: authorization.contractAddress,
-    r: authorization.r,
-    s: authorization.s,
-    chainId: toHex(authorization.chainId),
-    nonce: toHex(authorization.nonce),
-    ...(typeof authorization.yParity !== 'undefined'
-      ? { yParity: toHex(authorization.yParity) }
-      : {}),
-    ...(typeof authorization.v !== 'undefined' &&
-    typeof authorization.yParity === 'undefined'
-      ? { v: toHex(authorization.v) }
-      : {}),
-  }))
-
-const fmtRpcRequest = (request: SeismicTransactionRequest) => {
-  if (request.type === 'seismic') {
-    const { authorizationList, ...legacyCompatibleRequest } = request
-    const seismicFmt = formatTransactionRequest({
-      ...legacyCompatibleRequest,
-      type: 'legacy',
-    })
-    return {
-      ...seismicFmt,
-      type: SEISMIC_TX_TYPE,
-      ...(authorizationList
-        ? { authorizationList: formatAuthorizationList(authorizationList) }
-        : {}),
-    }
-  }
-
-  const { type, ...fmt } = formatTransactionRequest(request)
-  if (hasSeismicFields(request)) {
-    return { ...fmt, type: SEISMIC_TX_TYPE }
-  }
-  return { ...fmt, type }
-}
-
-/**
- * Chain formatters for Seismic transactions, providing formatting utilities for transaction requests.
- * @property {SeismicTransactionRequest} transactionRequest - Formatter configuration for transaction requests
- * @property {Function} transactionRequest.format - Formats a Seismic transaction request into the required RPC format
- * @param {SeismicTransactionRequest} request - The transaction request to be formatted
- * @returns {Object} A formatted transaction request object containing:
- *   - All properties from the formatted RPC request
- *   - `type` (optional) - Set to '0x4a' if encryption public key is present
- *   - `data` (optional) - Transaction data if present
- *   - `encryptionPubkey` (optional) - Public key for transaction encryption
- *   - `chainId` (optional) - Chain ID for the transaction
- * @remarks
- * This function is called by viem's call, estimateGas, and sendTransaction.
- * We can use this to parse transaction request before sending it to the node
- */
-export const seismicChainFormatters: ChainFormatters = {
-  transactionRequest: {
-    format: (request: SeismicTransactionRequest) => {
-      // @ts-expect-error: anvil requires chainId to be set but estimateGas doesn't set it
-      let chainId = request.chainId
-      const formattedRpcRequest = fmtRpcRequest(request)
-
-      if (request.type === 'seismic') {
-        if (!request.encryptionNonce) {
-          throw new Error(
-            'Encryption nonce is required for seismic transactions'
-          )
-        }
-        if (!request.encryptionPubkey) {
-          throw new Error(
-            'Encryption public key is required for seismic transactions'
-          )
-        }
-        if (request.messageVersion !== 0 && request.messageVersion !== 2) {
-          throw new Error(
-            'Message version must be set to 0 or 2 for seismic transactions'
-          )
-        }
-        if (!request.recentBlockHash) {
-          throw new Error(
-            'recentBlockHash is required for seismic transaction requests'
-          )
-        }
-        if (!request.expiresAtBlock) {
-          throw new Error(
-            'expiresAtBlock is required for seismic transaction requests'
-          )
-        }
-        if (request.signedRead === undefined) {
-          throw new Error(
-            'signedRead is required for seismic transaction requests'
-          )
-        }
-      }
-
-      const fmtSeismicReq = {
-        ...formattedRpcRequest,
-        chainId,
-        encryptionPubkey: request.encryptionPubkey,
-        encryptionNonce: request.encryptionNonce,
-        messageVersion: request.messageVersion,
-        recentBlockHash: request.recentBlockHash,
-        expiresAtBlock: request.expiresAtBlock,
-        signedRead: request.signedRead,
-      }
-      return fmtSeismicReq
-    },
-    type: 'transactionRequest',
-  },
-}
+import { seismicChainFormatters } from '@sviem/tx/seismicRpc.ts'
 
 export type CreateSeismicDevnetParams = {
   nodeHost: string
